:bug: UInteger should not derived from Integer  fix UInteger definition in nodeSet xml file ( according to OPCUA Spec part 5 v1.02 table 120) -  Integer and UInteger should derive from Number.  see https://opcfoundation-onlineapplications.org/mantis/view.php?id=3064

+++ test/misc/test_factories.js
@@ -248,7 +248,7 @@ describe("Factories: testing object factory", function () {
 
         factories.registerBasicType({
             name: "MyInteger",
-            subtype: "Integer",
+            subtype: "UInt16",
             defaultValue: 0
         });
